+++ radical_translations/core/management/commands/export.py
@@ -10,9 +10,9 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         self.stdout.write(
-            "Exporting Resources into CSV file resource.csv ...", ending=" "
+            "Exporting Resources into CSV file resources.csv ...", ending=" "
         )
-        resources = [r.to_dict() for r in Resource.objects.all()]
+        resources = [resource.to_dict() for resource in Resource.objects.all()]
 
         if not resources:
             self.stderr.write(self.style.NOTICE("No resources found!"))