+++ radical_translations/utils/models.py
@@ -4,6 +4,7 @@ from time import mktime, struct_time
 from typing import Dict, Optional
 
 from convertdate import french_republican
+from django.conf import settings
 from django.db import models
 from model_utils.models import TimeStampedModel
 
@@ -107,29 +108,25 @@ class Date(TimeStampedModel):
         return date
 
 
-def date_to_dict(date: Date) -> Dict:
+def date_to_dict(date: Date, label: Optional[str] = "date") -> Dict:
     if not date:
         return {
-            "date.date_display": "",
-            "date.date_display_classification": "",
-            "date.date_radical": "",
-            "date.date_radical_classification": "",
+            f"{label}.date_display": "",
+            f"{label}.date_display_classification": "",
+            f"{label}.date_radical": "",
+            f"{label}.date_radical_classification": "",
         }
 
     return {
-        "date.date_display": date.date_display,
-        "date.date_display_classification": get_controlled_terms_str(
+        f"{label}.date_display": date.date_display,
+        f"{label}.date_display_classification": get_controlled_terms_str(
             date.date_display_classification.all()
         ),
-        "date.date_radical": date.date_radical,
-        "date.date_radical_classification": get_controlled_terms_str(
+        f"{label}.date_radical": date.date_radical,
+        f"{label}.date_radical_classification": get_controlled_terms_str(
             date.date_radical_classification.all()
         ),
     }
-
-
-def get_controlled_terms_str(terms) -> str:
-    return "; ".join([ct.label for ct in terms])
 
 
 class EditorialClassificationModel(models.Model):
@@ -201,3 +198,37 @@ def get_geonames_place_from_gsx_place(name: str) -> Optional[Place]:
     address = matches.group("address").strip()
     country_code = matches.group("country_code")
     return Place.get_or_create_from_geonames(address, country_code=country_code)
+
+
+def get_controlled_terms_str(terms) -> str:
+    return f"{settings.EXPORT_MULTIVALUE_SEPARATOR} ".join(
+        [f"{ct.label} ({ct.vocabulary.label})" for ct in terms]
+    )
+
+
+def place_to_dict(place) -> Dict:
+    if not place:
+        return
+
+    return {
+        "geonames_id": place.geonames_id,
+        "address": place.address,
+        "class_description": place.class_description,
+        "country": place.country.name if place.country else "",
+        "feature_class": place.feature_class,
+        "lat": place.lat,
+        "lon": place.lon,
+    }
+
+
+def place_to_dict_value(place) -> str:
+    if not place:
+        return
+
+    if place.country:
+        return (
+            f"{place.geonames_id}{settings.EXPORT_FIELD_SEPARATOR}{place.address}"
+            f"{settings.EXPORT_FIELD_SEPARATOR}{place.country.name}"
+        )
+
+    return place.address