+++ CHANGELOG.rst
@@ -26,6 +26,10 @@ Added
 * Load analytics setting.
 * GitHub action for CI.
 * Command to export `Resources` to a CSV file.
+* Command to export `Organisations` and `Persons` to CSV files.
+* Command to export `Events` to a CSV file.
+* Docs: data model fields concordance.
+* Command to export `Places` to a CSV file.
 
 Changed
 ~~~~~~~